requirements.txt
streamlit
easyocr
pymupdf          
pillow
numpy
opencv-python-headless
st_img_pastebutton
openai
python-dotenv
torch
google-cloud-vision
google-cloud-documentai